Add ARToolKit v5
Loading ARToolKit and the initialization of it is done. Switching
between the Normal mode and the AR mode also works. "Viewport" went
through substantial modifications in order to provide a good experience
and smooth user interactions on mode switching.

It cannot display the camera image or perform marker tracking yet.

artoolkit.js is based on "artoolkit.api.js" from the "jsartoolkit" NPM
package, with some minor modifications to provide an access to the
created MediaStream object and to support iOS 11.